pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "hermes-financial"
version = "0.1.0"
description = "Multi-agent financial research framework"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [{ name = "Hermes Contributors" }]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Financial and Insurance Industry",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Office/Business :: Financial",
    "Typing :: Typed",
]
dependencies = [
    "llama-index-core>=0.12.0",
    "llama-index-llms-anthropic>=0.6.0",
    "llama-index-llms-openai>=0.3.0",
    "llama-index-embeddings-openai>=0.3.0",
    "llama-index-vector-stores-chroma>=0.4.0",
    "chromadb>=0.5.0",
    "httpx>=0.27.0",
    "pydantic>=2.0",
    "openpyxl>=3.1.0",
    "python-docx>=1.1.0",
    "matplotlib>=3.8.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    "edgartools>=3.0.0",
    "llama-index-llms-google-genai>=0.8.7",
]

[project.optional-dependencies]
llamaparse = ["llama-parse>=0.5.0"]
web = ["llama-index-readers-web>=0.3.0"]
pandas = ["pandas>=2.0.0"]
google = ["llama-index-llms-google-genai>=0.4.0"]
mistral = ["llama-index-llms-mistralai>=0.3.0"]
groq = ["llama-index-llms-groq>=0.3.0"]
ollama = ["llama-index-llms-ollama>=0.4.0"]
huggingface = ["llama-index-llms-huggingface-api>=0.3.0"]
cohere = ["llama-index-llms-cohere>=0.4.0"]
dev = [
    "pytest",
    "pytest-asyncio",
    "pytest-cov",
    "ruff",
    "mypy",
]
all = [
    "hermes-financial[llamaparse]",
    "hermes-financial[web]",
    "hermes-financial[pandas]",
    "hermes-financial[google]",
    "hermes-financial[mistral]",
    "hermes-financial[groq]",
    "hermes-financial[ollama]",
    "hermes-financial[huggingface]",
    "hermes-financial[cohere]",
    "hermes-financial[dev]",
]

[project.urls]
Homepage = "https://github.com/hermes-financial/hermes-financial"
Documentation = "https://github.com/hermes-financial/hermes-financial#readme"
Repository = "https://github.com/hermes-financial/hermes-financial"
Issues = "https://github.com/hermes-financial/hermes-financial/issues"

[tool.hatch.build.targets.wheel]
packages = ["hermes"]

[tool.ruff]
line-length = 99
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
